NFT related files
shop.html is the shop page and all the respective items are these item_ancient.html etc stuff, you can click the cards and be redirected into the individual pages and each individual pages have a return button to go back to the shop page. the buy button does not has a function yet hence left as "do not have function" alarm. return homepage on shop page is also not working yet as homepage has not been identified.